+++ src/IPGenerate/IPCatalogBuilder.cpp
@@ -42,12 +42,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <thread>
 
 #include "Compiler/Log.h"
-#include "Compiler/ProcessUtils.h"
 #include "Compiler/TclInterpreterHandler.h"
 #include "Compiler/WorkerThread.h"
 #include "IPGenerate/IPCatalogBuilder.h"
 #include "MainWindow/Session.h"
 #include "Utils/FileUtils.h"
+#include "Utils/ProcessUtils.h"
 #include "Utils/StringUtils.h"
 #include "nlohmann_json/json.hpp"
 using json = nlohmann::ordered_json;
